Implement improvements for Pro (#1502)

+++ core/src/main/java/org/jobrunr/jobs/filters/RetryFilter.java
@@ -63,8 +63,8 @@ public class RetryFilter implements ElectStateFilter {
         return getFailureCount(job) > getMaxNumberOfRetries(job);
     }
 
-    private long getFailureCount(Job job) {
-        return job.getJobStates().stream().filter(FAILED_STATES).count();
+    private int getFailureCount(Job job) {
+        return Math.toIntExact(job.getJobStates().stream().filter(FAILED_STATES).count());
     }
 
     private static boolean isJobNotFoundException(JobState newState) {
